+++ Assets/Scripts/Commands/COMF.cs
@@ -29,7 +29,7 @@ namespace Commands
             int result = ~memory[address];
 
             // Update Zero Flag
-            if (result == 0)
+            if (((byte)result) == 0)
             {
                 memory.ZeroFlag = 1;
             }
