+++ Assets/Scripts/Commands/INCF.cs
@@ -30,7 +30,7 @@ namespace Commands
             var result = memory[address] + 1;
 
             // Update Zero Flag
-            if (result == 0)
+            if (((byte)result) == 0)
             {
                 memory.ZeroFlag = 1;
             }
